CORE-7209 Move Delete btn to main tool list view
