+++ ui/de-lib/src/main/java/org/iplantc/de/admin/desktop/client/toolAdmin/ToolAdminView.java
@@ -166,7 +166,7 @@ public interface ToolAdminView extends IsWidget,
 
         String updateToolSuccessText();
 
-        String dialogWindowDeleteBtnText();
+        String deleteBtnText();
 
         String deleteToolSuccessText();
