+++ ui/de-lib/src/main/java/org/iplantc/de/theme/base/client/admin/toolAdmin/ToolAdminDisplayStrings.java
@@ -141,8 +141,8 @@ public interface ToolAdminDisplayStrings extends Messages {
     @Key("dialogWindowUpdateBtnText")
     String dialogWindowUpdateBtnText();
 
-    @Key("dialogWindowDeleteBtnText")
-    String dialogWindowDeleteBtnText();
+    @Key("deleteBtnText")
+    String deleteBtnText();
 
     @Key("deleteToolSuccessText")
     String deleteToolSuccessText();
