+++ ui/de-lib/src/main/java/org/iplantc/de/theme/base/client/admin/toolAdmin/ToolAdminDisplayStrings.properties
@@ -3,7 +3,7 @@ dialogWindowName = Tool Admin
 dialogWindowUpdateBtnText = Save/Update
 addToolSuccessText = Tool successfully added!
 updateToolSuccessText = Tool successfully updated!
-dialogWindowDeleteBtnText = Delete
+deleteBtnText = Delete
 deleteToolSuccessText = Tool successfully deleted!
 confirmOverwriteTitle = Public App Overwrite?
 confirmOverwriteDangerZone = Do not update container settings that are in use by tools in public apps unless it is certain the new container settings will not break reproducibility for those apps.\
